share code

+++ src/main/java/br/com/livelo/orderflight/domain/dtos/response/PriceResponseDTO.java
@@ -1,11 +1,13 @@
 package br.com.livelo.orderflight.domain.dtos.response;
 
+import java.math.BigDecimal;
+
 import lombok.Builder;
 import lombok.Data;
 
 @Data
 @Builder
 public class PriceResponseDTO {
-    public double amount;
-    public int pointsAmount;
+    public BigDecimal amount;
+    public BigDecimal pointsAmount;
 }
